Fix cast error in IterativeBinarySearchApp

+++ src/app/IterativeBinarySearchApp.java
@@ -11,7 +11,7 @@ public class IterativeBinarySearchApp {
         System.out.println("Enter the number you want to look for: ");
         Scanner in = FileUtils.loadFile("sortedInput.txt");
         int[] data = FileUtils.InputDataToArray(in);
-        int key = Math.random()*17+1;
+        int key = (int) (Math.random()*17+1);
         System.out.println("Reading data...");
         System.out.println("Looking for: "+key);
